Move matrix helpers into matrices directory
