+++ M2/Macaulay2/e/matrices/matrix-kbasis.cpp
