+++ M2/Macaulay2/e/matrices/matrix-ncbasis.hpp
